Refactor: 게임 관련 파일 분리, 게임 웹소켓 message 함수들 클래스로 묶기
