+++ FE/src/index.js
@@ -1,6 +1,6 @@
 import LoginPage from './pages/LoginPage.js';
 import PlayModePage from './pages/PlayModePage.js';
-import GamePage from './pages/GamePage.js';
+import GamePage from './game/GamePage.js';
 import RegisterPage from './pages/register/RegisterPage.js';
 
 import GameSettingPage from './pages/local/gameSetting/GameSettingPage.js';
@@ -18,9 +18,8 @@ import OnlineGameSettingDetailed from './pages/online/gameSetting/OnlineGameSett
 import OnlineGameSettingTournament from './pages/online/gameSetting/OnlineGameSettingTournament.js';
 
 import WaitingRoomPage from './pages/online/rooms/waitingRoom/WaitingRoomPage.js';
-import modifyGameSetting from './pages/online/rooms/ModifyGameSetting.js';
 
-import { GamewebsocketManager } from './websocket/GamewebsocketManager.js';
+import { GamewebsocketManager } from './game/GamewebsocketManager.js';
 import {
 	headerAddEventListeners,
 	profileButton
