+++ FE/src/pages/online/gameSetting/OnlineGameSettingPage.js
@@ -1,7 +1,7 @@
 import { changeUrl, changeUrlData, gamewsmanager } from '../../../index.js';
 import HorizontalButton from '../../../components/HorizontalButton.js';
 import VerticalButton from '../../../components/VerticalButton.js';
-import { Gamewebsocket } from '../../../websocket/Gamewebsocket.js';
+import { Gamewebsocket } from '../../../game/Gamewebsocket.js';
 import ButtonBackArrow from '../../../components/ButtonBackArrow.js';
 import { pongImage } from '../../../components/pongImage.js';
 import { createAndJoinRoom } from '../rooms/roomManager.js';
